services/requirements.txt
fastapi
uvicorn
pandas
numpy
scikit-learn
joblib
requests
python-dotenv
google-generativeai
prophet
